Downsampling can now operate over 2^31 - 1 (2 billion) labels.

+++ aloha-core/src/main/scala/com/eharmony/aloha/dataset/vw/multilabel/json/VwDownsampledMultilabeledJson.scala
@@ -7,14 +7,19 @@ import spray.json.{DefaultJsonProtocol, RootJsonFormat}
 import scala.collection.{immutable => sci}
 
 /**
-  * Created by ryan.deak on 11/6/17.
-  *
+  * JSON AST for `VwDownsampledMultilabelRowCreator`.
   * @param imports
   * @param features
   * @param namespaces
   * @param normalizeFeatures
-  * @param positiveLabels
-  * @param numDownsampledNegLabels
+  * @param positiveLabels string representing a function that will be used to extract positive
+  *                       labels from the input.
+  * @param numDownsampledNegLabels '''a positive value''' representing the number of negative
+  *                                labels to include in each row.  If this is less than the
+  *                                number of negative examples for a given row, then no
+  *                                downsampling of negatives will take place.
+  * @author deaktator
+  * @since 11/6/2017
   */
 final case class VwDownsampledMultilabeledJson(
     imports: sci.Seq[String],
@@ -22,7 +27,7 @@ final case class VwDownsampledMultilabeledJson(
     namespaces: Option[Seq[Namespace]] = Some(Nil),
     normalizeFeatures: Option[Boolean] = Some(false),
     positiveLabels: String,
-    numDownsampledNegLabels: Short
+    numDownsampledNegLabels: Int
 ) extends VwJsonLike {
 
   require(
